refactor: move register_readers to register
